+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## v0.3.13 2026-02-18
 ### Improvement
-- Remove async_load/unload from solar_charge.
+- Removed CALLBACK_HA_STARTED once triggered because unsubscribe is already done by HA.
+- Removed OCPP set current entity.
+- Removed async_load/unload from solar_charge.
 
 ## v0.3.12 2026-02-17
 ### Improvement