+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v0.3.13 2026-02-18
 ### Improvement
+- Log actual charging status.
 - Possible invalid OCPP state for setting charge current.
 - Removed CALLBACK_HA_STARTED once triggered because unsubscribe is already done by HA.
 - Removed OCPP set current entity.